Code quality, moved fixtures to conftest
